[Fix] 버그 수정

+++ src/main/java/com/goorm/cgcg/domain/Place.java
@@ -31,9 +31,9 @@ public class Place {
 
     private LocalTime time;
 
-    private Float latitude;
+    private double latitude;
 
-    private Float longitude;
+    private double longitude;
 
     private String name;
